Add multiple points in the JointTrajectory message
The aim of this change is to decrease the chances of timeouts when running it
on the actual hardware, where we seem to observe lags. The length (duration)
of the trajectory is specified in the trajectory/duration parameter.

A concern when having multiple points in the JointTrajectory message (spanning
over a certain duration) is that we now require a "stopping message" in order
to stop the trajectory execution. For example, if the duration is set to 5
seconds (which is pretty large; it should be at most several tenths of a
second), then the trajectory for the next 5 seconds is sent to the
manipulator. The manipulator will try to execute that trajectory, and if we
want it to stop we need to send a new message to override the 5-second
trajectory.