Change RobotModelLoader to a member variable
This fixes the warning message:

SEVERE WARNING!!! Attempting to unload library while objects created by this
loader exist in the heap! You should delete your objects before attempting to
unload the library or destroying the ClassLoader. The library will NOT be
unloaded.

This message is shown because the RobotModelLoader gets destroyed after the
JoggerTwist constructor.