Use directly the joint velocity for joint delta theta
The idea behind multiplying the joint angular velocities by the resolution
(i.e. time from start) is that the angular change in the joint angles can be
calculated using the relationship theta = omega * dt. However, it seems like
we're better off not multiplying the omega to calculate delta theta, observed
on the actual manipulator. We need to figure out why this is the case.