Calculate divergence between actual and commanded pose
The commanded pose is integrated to estimate the current joint states, as
opposed to using the JointState topic to get the current joint values. The
reason behind this is that the JointState topic has errors that add up, which
results in inaccurate end-effector pose. The topic also arrives slow, which
means that we can't take the moving average of several messages. To solve this
we assume that the end-effector follows exactly as we commanded, and keep
sending new commands based on the assumption that the previous command that we
sent has been achieved.